schedule pre made

+++ artists/models.py
@@ -152,7 +152,7 @@ class Rate(models.Model):
     artist_application = models.ForeignKey(ArtistApplication, on_delete=models.CASCADE, related_name='rates', null=True, blank=True)
     name = models.CharField(max_length=255)
     description = models.TextField(null=True, blank=True)
-    amount = models.PositiveIntegerField()
+    amount = models.DecimalField(max_digits=10, decimal_places=2)
 
     def __str__(self):
         return f'{self.artist}-{self.name}-{self.amount}'
